# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TriAIge Orchestrator — Fase 1 (Ingestão)
  version: "1.0"
  description: >
    Ingestão de documentos: criação de sessão, upload via URL pré-assinada,
    confirmação e finalização com publicação em Q2 para o MCP.
servers:
  - url: http://localhost:8080/api/orchestrator/v1

security:
  - bearerAuth: []

paths:
  /sessions:
    post:
      summary: Cria sessão de triagem + caso
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        '201':
          description: Sessão criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateSessionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'

  /sessions/{sessionId}/documents/presign:
    post:
      summary: Solicita URL de upload pré-assinada
      parameters:
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PresignDocumentRequest'
      responses:
        '201':
          description: URL gerada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresignDocumentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  /sessions/{sessionId}/documents/{documentId}/complete:
    post:
      summary: Confirma upload concluído
      parameters:
        - $ref: '#/components/parameters/SessionId'
        - $ref: '#/components/parameters/DocumentId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Documento marcado como RECEIVED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompleteDocumentResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  /sessions/{sessionId}/finalize:
    post:
      summary: Encerra recebimento e publica para o MCP (Q2)
      parameters:
        - $ref: '#/components/parameters/SessionId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Sessão finalizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinalizeSessionResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  /sessions/{sessionId}:
    get:
      summary: Consulta status da sessão e documentos
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Detalhe da sessão
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  /sessions/{sessionId}/mcp-result:
    post:
      summary: Callback do triaige-srv-mcp-ai (Fase 2) com o resultado do pipeline
      description: >
        Rede interna apenas (não exposto via ALB/API Gateway em produção). Autenticado por
        X-Internal-Token, não por Bearer — não faz parte de security: [bearerAuth].
      security:
        - internalToken: []
      parameters:
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/McpResultCallbackRequest'
      responses:
        '200':
          description: Resultado aplicado
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    internalToken:
      type: apiKey
      in: header
      name: X-Internal-Token

  parameters:
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    DocumentId:
      name: documentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid

  responses:
    BadRequest:
      description: Payload inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Token ausente, inválido, expirado ou credencial inativa
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflito de estado ou de idempotência
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        code:
          type: string
        error:
          type: string
        message:
          type: string
        path:
          type: string

    CreateSessionRequest:
      type: object
      required: [caso]
      properties:
        caso:
          type: object
          required: [titulo, areaJuridica, tipoCaso]
          properties:
            titulo:
              type: string
              maxLength: 255
            areaJuridica:
              type: string
              maxLength: 30
            tipoCaso:
              type: string
              maxLength: 50

    CreateSessionResponse:
      type: object
      properties:
        sessionId:
          type: string
          format: uuid
        protocolo:
          type: string
        correlationId:
          type: string
          format: uuid
        status:
          type: string
          enum: [RECEIVING_DOCUMENTS]
        createdAt:
          type: string
          format: date-time

    PresignDocumentRequest:
      type: object
      required: [nomeArquivoOriginal, tipoDocumento, contentType, tamanhoBytesEstimado]
      properties:
        attachmentGroupId:
          type: string
          format: uuid
        partNumber:
          type: integer
          minimum: 1
          default: 1
        nomeArquivoOriginal:
          type: string
          maxLength: 255
        tipoDocumento:
          type: string
          maxLength: 30
        contentType:
          type: string
          maxLength: 100
        tamanhoBytesEstimado:
          type: integer
          format: int64

    PresignDocumentResponse:
      type: object
      properties:
        documentId:
          type: string
          format: uuid
        attachmentGroupId:
          type: string
          format: uuid
        partNumber:
          type: integer
        uploadUrl:
          type: string
        uploadUrlExpiresAt:
          type: string
          format: date-time
        rawBucket:
          type: string
        rawObjectKey:
          type: string

    CompleteDocumentResponse:
      type: object
      properties:
        documentId:
          type: string
          format: uuid
        status:
          type: string
          enum: [RECEIVED]
        tamanhoBytes:
          type: integer
          format: int64

    FinalizeSessionResponse:
      type: object
      properties:
        sessionId:
          type: string
          format: uuid
        protocolo:
          type: string
        status:
          type: string
          enum: [QUEUED_FOR_PROCESSING]
        documentsCount:
          type: integer

    SessionDetailResponse:
      type: object
      properties:
        sessionId:
          type: string
          format: uuid
        protocolo:
          type: string
        status:
          type: string
          enum: [RECEIVING_DOCUMENTS, QUEUED_FOR_PROCESSING, CANCELLED, COMPLETED, PARTIALLY_COMPLETED, FAILED]
        documents:
          type: array
          items:
            type: object
            properties:
              documentId:
                type: string
                format: uuid
              status:
                type: string
                enum: [PENDING_UPLOAD, RECEIVED, QUEUED, UPLOAD_FAILED, OCR_IN_PROGRESS, OCR_DONE, ANONYMIZED, GROUPED, TRUSTED, OCR_FAILED, PROCESSING_FAILED]
              nomeArquivoOriginal:
                type: string

    McpResultCallbackRequest:
      type: object
      required: [sessionId, status]
      description: Espelha McpResultCallbackPayload do triaige-srv-mcp-ai (spec da Fase 2, seção 6).
      properties:
        schemaVersion:
          type: string
          default: "1.0"
        sessionId:
          type: string
          format: uuid
        correlationId:
          type: string
          format: uuid
        status:
          type: string
          enum: [COMPLETED, PARTIALLY_COMPLETED, FAILED]
        processedGroups:
          type: array
          items:
            type: object
            properties:
              attachmentGroupId:
                type: string
                format: uuid
              trustedBucket:
                type: string
              trustedObjectKey:
                type: string
              tipoDocumento:
                type: string
              resumido:
                type: boolean
              piiRedactedCategories:
                type: array
                items:
                  type: string
        failedDocuments:
          type: array
          items:
            type: object
            properties:
              documentId:
                type: string
                format: uuid
              errorMessage:
                type: string
        completedAt:
          type: string
          format: date-time