# src/main/java/br/com/triaige/orchestrator/api/dto/response/CompleteDocumentResponse.java
package br.com.triaige.orchestrator.api.dto.response;

import br.com.triaige.orchestrator.domain.enums.DocumentStatus;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.UUID;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class CompleteDocumentResponse {

    private UUID documentId;
    private DocumentStatus status;
    private Long tamanhoBytes;
}