# src/main/java/br/com/triaige/orchestrator/api/dto/response/SessionDetailResponse.java
package br.com.triaige.orchestrator.api.dto.response;

import br.com.triaige.orchestrator.domain.enums.DocumentStatus;
import br.com.triaige.orchestrator.domain.enums.SessionStatus;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.List;
import java.util.UUID;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class SessionDetailResponse {

    private UUID sessionId;
    private String protocolo;
    private SessionStatus status;
    private List<DocumentSummary> documents;

    @Data
    @Builder
    @NoArgsConstructor
    @AllArgsConstructor
    public static class DocumentSummary {
        private UUID documentId;
        private DocumentStatus status;
        private String nomeArquivoOriginal;
    }
}